pyproject.toml
[build-system]
requires = ["setuptools>=77", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "delfin-complat"
version = "1.0.8"
description = "Automated ORCA, xTB and CREST workflow for predicting preferred spin states and associated redox potentials."
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.10"
authors = [{ name = "Maximilian Hartmann" }]
license = "LGPL-3.0-or-later"
urls = { "Homepage" = "https://github.com/ComPlat/DELFIN", "Repository" = "https://github.com/ComPlat/DELFIN" }
dependencies = [
    "mendeleev>=0.14",
    "psutil",
    "ase>=3.22.0",
    "numpy>=1.19.5",
    "scipy>=1.6.0",
    "matplotlib>=3.3.4",
    "rdkit>=2023.3.1",
    "openbabel-wheel>=3.1.0",
    "python-docx>=0.8.11",
    "pymol-open-source>=3.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov",
    "black",
    "flake8",
    "ruff",
]
ai = [
    "anthropic>=0.18.0",
    "openai>=1.12.0",
    "ollama>=0.1.0",
    "jsonschema>=4.0.0",
]

[project.scripts]
delfin = "delfin.main:main"
delfin_ESD = "delfin.cli_esd_report:main"
delfin_IR = "delfin.cli_ir_report:main"
delfin-json = "delfin.cli_delfin_collect:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["delfin*"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
]
ignore = [
    "E501", # handled by black (line length)
]

[tool.ruff.lint.isort]
known-first-party = ["delfin"]
